vertex_ai_requirements.txt
# Google Cloud和Vertex AI依赖
google-cloud-aiplatform>=1.38.0
# google-cloud-sql-connector>=1.0.0  # 使用 cloud-sql-python-connector 替代
google-cloud-storage>=2.10.0
google-cloud-logging>=3.8.0

# 数据库连接器
PyMySQL>=1.1.0
SQLAlchemy>=2.0.0
cloud-sql-python-connector>=1.4.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 异步处理
asyncio>=3.4.3
aiofiles>=23.2.0
asyncpg>=0.28.0

# 环境变量和配置
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# 日志和错误处理
structlog>=23.1.0
rich>=13.5.0
tenacity>=8.2.0

# HTTP客户端
httpx>=0.25.0
requests>=2.31.0

# 测试依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# 开发工具
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.4.0

# 其他实用工具
tqdm>=4.66.0
click>=8.1.0
tabulate>=0.9.0
colorama>=0.4.6 